Moved our common.ts script to the parent folder.
